requirements.txt
langchain_community
langchain_google_genai
python-dotenv
pydantic
langchain_huggingface
sentence-transformers
scikit-learn
faiss-cpu
bs4
langchain-experimental